+++ user_guide_src/source/changelog.rst
@@ -12,6 +12,7 @@ Release Date: Not Released
    -  Changed ``CI_Log`` to append ``PHP_EOL`` instead of ``\n`` at the end of log messages.
    -  Improved performance in :doc:`Cache Library <libraries/caching>` 'redis' driver with non-scalar variables.
    -  Altered the :doc:`Session Library <libraries/session>` 'files' driver to log error and trigger a session start failure instead of throwing an ``Exception`` in case of unusable ``$config['sess_save_path']``.
+   -  Updated the :doc:`Session <libraries/session>` and :doc:`Cache <libraries/caching>` libraries' 'redis' driver to work with phpRedis 5.
 
 Bug fixes for 3.1.11
 ====================